+++ resources/views/auth/reset-password.blade.php
@@ -18,7 +18,7 @@
                 </div>
                 <p>{{__('cms::auth.pages.new_password.text')}}</p>
                 <div class="mt-10">
-                    <a href="{{route('cms.backend.auth.login')}}"
+                    <a href="{{route('cms.auth.login')}}"
                        class="group relative w-full flex justify-center py-2 px-4 border border-transparent text-sm font-medium rounded-md text-white bg-indigo-600 hover:bg-indigo-500 focus:outline-none focus:border-indigo-700 focus:shadow-outline-indigo active:bg-indigo-700 transition duration-150 ease-in-out">
                         {{__('cms::auth.pages.new_password.goto')}}
