+++ src/Http/Controllers/Backend/UsersController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Bambamboole\LaravelCms\Http\Controllers;
+namespace Bambamboole\LaravelCms\Http\Controllers\Backend;
 
 use Bambamboole\LaravelCms\Http\Requests\CreateUserRequest;
 use Bambamboole\LaravelCms\Http\Requests\UpdateUserRequest;
