+++ src/Services/CmsRouter.php
@@ -5,11 +5,11 @@ namespace Bambamboole\LaravelCms\Services;
 use Bambamboole\LaravelCms\Http\Controllers\Auth\ForgotPasswordController;
 use Bambamboole\LaravelCms\Http\Controllers\Auth\LoginController;
 use Bambamboole\LaravelCms\Http\Controllers\BlogController;
-use Bambamboole\LaravelCms\Http\Controllers\MenusController;
-use Bambamboole\LaravelCms\Http\Controllers\PagesController;
-use Bambamboole\LaravelCms\Http\Controllers\PostsController;
-use Bambamboole\LaravelCms\Http\Controllers\ProfileController;
-use Bambamboole\LaravelCms\Http\Controllers\UsersController;
+use Bambamboole\LaravelCms\Http\Controllers\Backend\MenusController;
+use Bambamboole\LaravelCms\Http\Controllers\Backend\PagesController;
+use Bambamboole\LaravelCms\Http\Controllers\Backend\PostsController;
+use Bambamboole\LaravelCms\Http\Controllers\Backend\ProfileController;
+use Bambamboole\LaravelCms\Http\Controllers\Backend\UsersController;
 use Bambamboole\LaravelCms\Http\Middleware\Authenticate;
 use Bambamboole\LaravelCms\Http\Middleware\SetInertiaConfiguration;
 use Bambamboole\LaravelCms\Http\Middleware\SetLocale;
@@ -64,10 +64,10 @@ class CmsRouter
             });
     }
 
-    public function registerBackendRoutes(): void
+    public function registerAuthRoutes()
     {
         $this->router->middleware([$this->config->get('cms.backend.middleware'), SetLocale::class])
-            ->as('cms.backend.')
+            ->as('cms.')
             ->prefix($this->config->get('cms.backend.url'))
             ->group(function (Router $router) {
                 $router->get('/login', [LoginController::class, 'showLoginForm'])->name('auth.login');
@@ -78,7 +78,10 @@ class CmsRouter
                 $router->post('/password/forgot', [ForgotPasswordController::class, 'sendResetLinkEmail'])->name('password.email');
                 $router->get('/password/reset/{token}', [ForgotPasswordController::class, 'showNewPassword'])->name('password.reset');
             });
+    }
 
+    public function registerBackendRoutes(): void
+    {
         $this->router
             ->middleware([
                 $this->config->get('cms.backend.middleware'),
